requirements.txt
pandas==1.3.1
torch==1.9.0
tqdm==4.61.2
open3d==0.13.0
pydicom==2.1.2
imageio==2.9.0
torchvision==0.10.0
numpy==1.21.1
easydict==1.9
Pillow==8.3.1
scikit_learn==0.24.2
